autopal/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic-settings==2.2.1
